+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Package `wal-prelude` now provides all functionality to bootstrap
+  the configuration. Both the provided `init.el` template and
+  `wal-setup-ci.el` use it.
+
 ### Changed
 
 - All `use-package` definitions now use a more relaxed style, leaving